+++ integration-tests/__tests__/find_related_files.test.js
@@ -13,7 +13,7 @@ import runJest from '../runJest';
 import os from 'os';
 import path from 'path';
 
-const {cleanup, writeFiles} = require('../utils');
+const {cleanup, writeFiles} = require('../Utils');
 
 const SkipOnWindows = require('../../scripts/SkipOnWindows');
 const DIR = path.resolve(os.tmpdir(), 'find_related_tests_test');