+++ integration-tests/__tests__/jest.config.js.test.js
@@ -11,7 +11,7 @@
 
 const path = require('path');
 const SkipOnWindows = require('../../scripts/SkipOnWindows');
-const {extractSummary, cleanup, writeFiles} = require('../utils');
+const {extractSummary, cleanup, writeFiles} = require('../Utils');
 const runJest = require('../runJest');
 
 const DIR = path.resolve(__dirname, '../jest.config.js');